heroku/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cmdstanpy==0.4.0
convertdate==2.2.1
cycler==0.10.0
Cython==0.29.20
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
ephem==3.7.7.1
fbprophet==0.6
gunicorn==19.7.1
holidays==0.10.2
idna==2.9
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.2.2
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
setuptools-git==1.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
